Add StoreService
